+++ episodic/orchestration/_checkpoint_dto.py
@@ -1,16 +1,14 @@
 """Checkpoint DTOs for resumable generation orchestration."""
 
 import dataclasses as dc
-import typing as typ
+import datetime as dt  # noqa: TC003 - runtime annotation inspection needs this name.
+import json
 
-from ._dto import (
+from ._payload_dto import (
     ActionExecutionResult,
     _normalize_non_empty_text,
     _normalize_string_fields,
 )
-
-if typ.TYPE_CHECKING:
-    import datetime as dt
 
 
 @dc.dataclass(frozen=True, slots=True)
@@ -49,6 +47,11 @@ class WorkflowCheckpoint:
         if not isinstance(self.payload, dict):
             msg = "payload must be a mapping object."
             raise TypeError(msg)
+        try:
+            json.dumps(self.payload, allow_nan=False)
+        except (TypeError, ValueError) as exc:
+            msg = "payload must be JSON-serializable."
+            raise TypeError(msg) from exc
         object.__setattr__(self, "payload", dict(self.payload))
