+++ episodic/orchestration/_result_dto.py
@@ -3,7 +3,7 @@
 import dataclasses as dc
 import typing as typ
 
-from ._dto import ActionExecutionResult, ExecutionPlan
+from ._payload_dto import ActionExecutionResult, ExecutionPlan
 
 if typ.TYPE_CHECKING:
     from episodic.llm import LLMUsage
